ISU-AI-Support/requirements-core.txt
neo4j
openai
python-dotenv
fastapi
python-multipart
uvicorn[standard]
requests
nest_asyncio
mongoengine
cloudinary
PyJWT
pika
